python/requirements.txt
google-cloud-pubsub==0.38.0